wo#7091 . make it possible to access ikev2 microcode table externally, for unit tests
